+++ src/engine/r_things.c
@@ -33,6 +33,7 @@
 #include "r_clipper.h"
 #include "m_misc.h"
 #include "con_console.h"
+#include "i_swap.h"
 
 #include <stdlib.h>
 
@@ -599,7 +600,7 @@ void R_DrawPSprite(pspdef_t* psp, sector_t* sector, player_t* player) {
 		color = D_RGBA(255, 255, 255, alpha);
 	}
 	else {
-		color = R_GetSectorLight(alpha, sector->colors[LIGHT_THING]);
+		color = LONG(R_GetSectorLight(alpha, sector->colors[LIGHT_THING]));
 	}
 
 	spritenum = sprframe->lump[0];
